pkgdown/extra.css

Plot VILMA Regionalization Rasters

Provides a quick visualization of a vilma.region object produced by vilma.regionalize. The plot shows two maps in a vertical layout: (i) the raster of region (group) IDs and (ii) the raster of mean within-region phylogenetic dissimilarity (mean PD distance) per cell.

Usage

# S3 method for class 'vilma.region'
plot(x, ...)

Arguments

x

An object of class vilma.region, typically generated by vilma.regionalize.

...

Additional arguments (currently not used).

Value

Produces raster plots for the region IDs and mean within-region dissimilarity. Invisibly returns x.

Details

This method is intended for quick inspection of regionalization results. Internally it uses plot to display the rasters stored in x$raster. A small helper is included to optionally label cells with numeric values when the raster contains finite values.

See also

Author

Omar Daniel Leon-Alvarado https://leon-alvarado.weebly.com/ J. Angel Soto-Centeno https://www.mormoops.com/